pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "self-healing-ml"
version = "1.0.0"
description = "Self-Healing ML Pipelines"
dependencies = [
    "numpy>=2.0.0",
    "pandas>=2.0.0", 
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "rich>=13.0.0",
    "prometheus-client>=0.17.0",
    "joblib>=1.3.0",
]
